+++ sdk/src/client/api/mod.rs
@@ -1,7 +1,11 @@
 mod add;
+mod connect;
 mod remove;
 mod remove_all;
+mod try_connect;
 
 pub use self::add::*;
+pub use self::connect::*;
 pub use self::remove::*;
 pub use self::remove_all::*;
+pub use self::try_connect::*;