+++ sdk/src/pool/mod.rs
@@ -6,7 +6,7 @@
 
 use std::borrow::Cow;
 use std::collections::{HashMap, HashSet};
-use std::future::Future;
+use std::future::{Future, IntoFuture};
 use std::iter::Zip;
 use std::pin::Pin;
 use std::sync::atomic::Ordering;
@@ -352,19 +352,7 @@ impl RelayPool {
         self.inner.remove_all_relays(force).await
     }
 
-    /// Connect to all added relays
-    ///
-    /// Attempts to initiate a connection for every relay currently in
-    /// [`RelayStatus::Initialized`] or [`RelayStatus::Terminated`].
-    /// A background connection task is spawned for each such relay, which then tries
-    /// to establish the connection.
-    /// Any relay not in one of these two statuses is skipped.
-    ///
-    /// For further details, see the documentation of [`Relay::connect`].
-    ///
-    /// [`RelayStatus::Initialized`]: crate::relay::RelayStatus::Initialized
-    /// [`RelayStatus::Terminated`]: crate::relay::RelayStatus::Terminated
-    pub async fn connect(&self) {
+    pub(crate) async fn connect(&self) {
         // Lock with read shared access
         let relays = self.inner.atomic.relays.read().await;
 
@@ -374,11 +362,7 @@ impl RelayPool {
         }
     }
 
-    /// Waits for relays connections
-    ///
-    /// Wait for relays connections at most for the specified `timeout`.
-    /// The code continues when the relays are connected or the `timeout` is reached.
-    pub async fn wait_for_connection(&self, timeout: Duration) {
+    pub(crate) async fn wait_for_connection(&self, timeout: Duration) {
         // Lock with read shared access
         let relays = self.inner.atomic.relays.read().await;
 
@@ -392,20 +376,7 @@ impl RelayPool {
         future::join_all(futures).await;
     }
 
-    /// Try to establish a connection with the relays.
-    ///
-    /// Attempts to establish a connection for every relay currently in
-    /// [`RelayStatus::Initialized`] or [`RelayStatus::Terminated`]
-    /// without spawning the connection task if it fails.
-    /// This means that if the connection fails, no automatic retries are scheduled.
-    /// Use [`RelayPool::connect`] if you want to immediately spawn a connection task,
-    /// regardless of whether the initial connection succeeds.
-    ///
-    /// For further details, see the documentation of [`Relay::try_connect`].
-    ///
-    /// [`RelayStatus::Initialized`]: crate::relay::RelayStatus::Initialized
-    /// [`RelayStatus::Terminated`]: crate::relay::RelayStatus::Terminated
-    pub async fn try_connect(&self, timeout: Duration) -> Output<()> {
+    pub(crate) async fn try_connect(&self, timeout: Duration) -> Output<()> {
         // Lock with read shared access
         let relays = self.inner.atomic.relays.read().await;
 
@@ -416,7 +387,7 @@ impl RelayPool {
         // Filter only relays that can connect and compose futures
         for relay in relays.values().filter(|r| r.status().can_connect()) {
             urls.push(relay.url().clone());
-            futures.push(relay.try_connect(timeout));
+            futures.push(relay.try_connect().timeout(timeout).into_future());
         }
 
         // TODO: use semaphore to limit number concurrent connections?
@@ -491,7 +462,7 @@ impl RelayPool {
         let relay: &Relay = self.internal_relay(&relays, &url)?;
 
         // Try to connect
-        relay.try_connect(timeout).await?;
+        relay.try_connect().timeout(timeout).await?;
 
         Ok(())
     }